fix: add account button
